refactor: split media's components (#112)
add a new hook useColorDarkenLighten to reduce repeating code
